+++ tests/testthat/test-check_convergence.R
@@ -16,8 +16,13 @@ test_that("check_convergence", {
   expect_true(check_convergence(model))
   expect_equal(
     check_convergence(model),
-    structure(TRUE, gradient = 0.000280307452338331),
+    structure(TRUE, gradient = NA_real_),
     tolerance = 1e-3
+  )
+  expect_warning(
+    check_convergence(x = model),
+    regex = "Argument `x` is deprecated",
+    fixed = TRUE
   )
 })